[Minor] Move around code and rename wasmfs.cpp -> syscalls.cpp (#15391)
- Rename wasmfs.cpp -> syscalls.cpp
- New streams.h file
- Moved some functions to file.cpp and file.h
