+++ system/lib/wasmfs/file.cpp
@@ -7,6 +7,8 @@
 // See https://github.com/emscripten-core/emscripten/issues/15041.
 
 #include "file.h"
+// TODO remove this in the next PR.
+#include "file_table.h"
 
 namespace wasmfs {
 //
@@ -38,4 +40,76 @@ __wasi_errno_t MemoryFile::read(uint8_t* buf, size_t len, off_t offset) {
 
   return __WASI_ERRNO_SUCCESS;
 };
+//
+// Path Parsing utilities
+//
+std::shared_ptr<Directory> getDir(std::vector<std::string>::iterator begin,
+                                  std::vector<std::string>::iterator end,
+                                  long& err) {
+
+  std::shared_ptr<File> curr;
+  // Check if the first path element is '/', indicating an absolute path.
+  if (*begin == "/") {
+    curr = getRootDirectory();
+    begin++;
+  }
+
+  for (auto it = begin; it != end; ++it) {
+    auto directory = curr->dynCast<Directory>();
+
+    // If file is nullptr, then the file was not a Directory.
+    // TODO: Change this to accommodate symlinks
+    if (!directory) {
+      err = -ENOTDIR;
+      return nullptr;
+    }
+
+    // Find the next entry in the current directory entry
+#ifdef WASMFS_DEBUG
+    directory->locked().printKeys();
+#endif
+    curr = directory->locked().getEntry(*it);
+
+    // Requested entry (file or directory)
+    if (!curr) {
+      err = -ENOENT;
+      return nullptr;
+    }
+
+#ifdef WASMFS_DEBUG
+    emscripten_console_log(it->c_str());
+#endif
+  }
+
+  auto currDirectory = curr->dynCast<Directory>();
+
+  if (!currDirectory) {
+    err = -ENOTDIR;
+    return nullptr;
+  }
+
+  return currDirectory;
+}
+
+std::vector<std::string> splitPath(char* pathname) {
+  std::vector<std::string> pathParts;
+  char newPathName[strlen(pathname) + 1];
+  strcpy(newPathName, pathname);
+
+  // TODO: Support relative paths. i.e. specify cwd if path is relative.
+  // TODO: Other path parsing edge cases.
+  char* current;
+  // Handle absolute path.
+  if (newPathName[0] == '/') {
+    pathParts.push_back("/");
+  }
+
+  current = strtok(newPathName, "/");
+  while (current != NULL) {
+    pathParts.push_back(current);
+    current = strtok(NULL, "/");
+  }
+
+  return pathParts;
+}
 } // namespace wasmfs
