+++ system/lib/wasmfs/file_table.h
@@ -29,16 +29,6 @@ template<typename T> bool addWillOverFlow(T a, T b) {
   return false;
 }
 
-// Obtains parent directory of a given pathname.
-// Will return a nullptr if the parent is not a directory.
-std::shared_ptr<Directory> getDir(std::vector<std::string>::iterator begin,
-                                  std::vector<std::string>::iterator end,
-                                  long& err);
-
-// Return a vector of the '/'-delimited components of a path. The first element
-// will be "/" iff the path is an absolute path.
-std::vector<std::string> splitPath(char* pathname);
-
 // Access mode, file creation and file status flags for open.
 using oflags_t = uint32_t;
